add debug to smoother
