+++ RecoTracker/TrackProducer/src/TrackProducerAlgorithm.cc
@@ -30,6 +30,8 @@
 #include "TrackingTools/TrackFitters/interface/RecHitSorter.h"
 #include "DataFormats/TrackReco/interface/TrackBase.h"
 
+#include<sstream>
+
 // #define VI_DEBUG
 // #define STAT_TSB
 
@@ -250,33 +252,36 @@ TrackProducerAlgorithm<reco::GsfTrack>::buildTrack (const TrajectoryFitter * the
   
   auto theTraj = new Trajectory( std::move(trajTmp) );
   theTraj->setSeedRef(seedRef);
-  
-  //  TrajectoryStateOnSurface innertsos;
-  // TrajectoryStateOnSurface outertsos;
 
-  // if (theTraj->direction() == alongMomentum) {
-  //  innertsos = theTraj->firstMeasurement().updatedState();
-  //  outertsos = theTraj->lastMeasurement().updatedState();
-  // } else { 
-  //  innertsos = theTraj->lastMeasurement().updatedState();
-  //  outertsos = theTraj->firstMeasurement().updatedState();
-  // }
-  //     std::cout
-  //       << "Nr. of first / last states = "
-  //       << innertsos.components().size() << " "
-  //       << outertsos.components().size() << std::endl;
-  //     std::vector<TrajectoryStateOnSurface> components = 
-  //       innertsos.components();
-  //     double sinTheta = 
-  //       sin(innertsos.globalMomentum().theta());
-  //     for ( std::vector<TrajectoryStateOnSurface>::const_iterator ic=components.begin();
-  // 	  ic!=components.end(); ic++ ) {
-  //       std::cout << " comp " << ic-components.begin() << " "
-  // 		<< (*ic).weight() << " "
-  // 		<< (*ic).localParameters().vector()[0]/sinTheta << " "
-  // 		<< sqrt((*ic).localError().matrix()[0][0])/sinTheta << std::endl;
-  //     }
-  
+#ifdef EDM_ML_DEBUG  
+  TrajectoryStateOnSurface innertsos;
+  TrajectoryStateOnSurface outertsos;
+
+  if (theTraj->direction() == alongMomentum) {
+    innertsos = theTraj->firstMeasurement().updatedState();
+    outertsos = theTraj->lastMeasurement().updatedState();
+  } else { 
+    innertsos = theTraj->lastMeasurement().updatedState();
+     outertsos = theTraj->firstMeasurement().updatedState();
+  }
+  std::ostringstream ss;
+  auto dc = [&](TrajectoryStateOnSurface const & tsos){ 
+     std::vector<TrajectoryStateOnSurface> const & components = tsos.components();
+     auto sinTheta =  std::sin(tsos.globalMomentum().theta());
+     for (auto const & ic : components) ss << ic.weight() << "/"; ss << "\n";
+     for (auto const & ic : components) ss << ic.localParameters().vector()[0]/sinTheta << "/"; ss << "\n";
+     for (auto const & ic : components) ss << std::sqrt(ic.localError().matrix()(0,0))/sinTheta << "/"; 
+  };
+  ss  << "\ninner comps\n";
+  dc(innertsos);
+  ss  << "\nouter comps\n";
+  dc(outertsos);
+  LogDebug("TrackProducer")
+ 	   << "Nr. of first / last states = "
+  	   << innertsos.components().size() << " "
+           << outertsos.components().size() << ss.str();
+#endif  
+
   ndof = 0;
   for (auto const & tm : theTraj->measurements()) {
     auto const & h = tm.recHitR();
@@ -335,7 +340,6 @@ TrackProducerAlgorithm<reco::GsfTrack>::buildTrack (const TrajectoryFitter * the
   math::XYZVector mom( p.x(), p.y(), p.z() );
   
   LogDebug("GsfTrackProducer") << "pos=" << v << " mom=" << p << " pt=" << p.perp() << " mag=" << p.mag();
-  std::cout << "GsfTrackProducer " << "pos=" << v << " mom=" << p << " pt=" << p.perp() << " mag=" << p.mag() << std::endl;
   
   auto theTrack = new reco::GsfTrack(theTraj->chiSquared(),
 				int(ndof),//FIXME fix weight() in TrackingRecHit
