+++ TrackingTools/GsfTracking/src/GsfTrajectorySmoother.cc
@@ -6,6 +6,9 @@
 #include "TrackingTools/TrackFitters/interface/TrajectoryStateCombiner.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 
+#include "DebugHelpers.h"
+
+
 GsfTrajectorySmoother::GsfTrajectorySmoother(const GsfPropagatorWithMaterial& aPropagator,
 					     const TrajectoryStateUpdator& aUpdator,
 					     const MeasurementEstimator& aEstimator,
@@ -37,8 +40,6 @@ GsfTrajectorySmoother::GsfTrajectorySmoother(const GsfPropagatorWithMaterial& aP
 
   if(!theGeometry) theGeometry = &dummyGeometry;
 
-  //   static SimpleConfigurable<bool> timeConf(false,"GsfTrajectorySmoother:activateTiming");
-  //   theTiming = timeConf.value();
 }
 
 GsfTrajectorySmoother::~GsfTrajectorySmoother() {
@@ -54,15 +55,6 @@ GsfTrajectorySmoother::~GsfTrajectorySmoother() {
 Trajectory 
 GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
 
-  //   static TimingReport::Item* propTimer =
-  //     &(*TimingReport::current())[string("GsfTrajectorySmoother:propagation")];
-  //   propTimer->switchCPU(false);
-  //   if ( !theTiming )  propTimer->switchOn(false);
-  //   static TimingReport::Item* updateTimer =
-  //     &(*TimingReport::current())[string("GsfTrajectorySmoother:update")];
-  //   updateTimer->switchCPU(false);
-  //   if ( !theTiming )  updateTimer->switchOn(false);
-  
   if(aTraj.empty()) return Trajectory();
   
   const Propagator* usePropagator = theAlongPropagator;
@@ -81,7 +73,7 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
   predTsos.rescaleError(theErrorRescaling);
 
   if(!predTsos.isValid()) {
-    edm::LogInfo("GsfTrajectorySmoother") 
+    edm::LogInfo("GsfTrackFitters") 
       << "GsfTrajectorySmoother: predicted tsos of last measurement not valid!";
     return Trajectory();
   }
@@ -128,49 +120,30 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
   
   TrajectoryStateCombiner combiner;
 
+  int hitcounter = avtm.size()-1;
   for(std::vector<TM>::const_reverse_iterator itm = avtm.rbegin() + 1; 
       itm < avtm.rend() - 1; ++itm) {
-    {
-      //       TimeMe t(*propTimer,false);
-      //       //
-      //       // check type of surface in case of invalid hit
-      //       // (in this version only propagations to planes are
-      //       // supported for multi trajectory states)
-      //       //
-      //       if ( !(*itm).recHit().isValid() ) {
-      // 	const BoundPlane* plane = 
-      // 	  dynamic_cast<const BoundPlane*>(&(*itm).recHit().det().surface());
-      // 	//
-      // 	// no plane: insert invalid 
-      // 	if ( plane==0 ) {
-      // 	  myTraj.push(TM(TrajectoryStateOnSurface(),
-      // 			 (*itm).recHit());
-      // 	  continue;
-      // 	}
-      //     }
-      predTsos = usePropagator->propagate(currTsos,
+
+     predTsos = usePropagator->propagate(currTsos,
 					 *(*itm).recHit()->surface());
-    }
     if ( predTsos.isValid() && theConvolutor && theMatBeforeUpdate )
       predTsos = (*theConvolutor)(predTsos,
 				  usePropagator->propagationDirection());
     if(!predTsos.isValid()) {
-      edm::LogInfo("GsfTrajectorySmoother") << "GsfTrajectorySmoother: predicted tsos not valid!";
+      edm::LogInfo("GsfTrackFitters") << "GsfTrajectorySmoother: predicted tsos not valid!";
       return Trajectory();
     }
     if ( theMerger )  predTsos = theMerger->merge(predTsos);
 
     if((*itm).recHit()->isValid()) {
       //update
-      {
-	// 	TimeMe t(*updateTimer,false);
 	currTsos = updator()->update(predTsos, *(*itm).recHit());
-      }
+
       if ( currTsos.isValid() && theConvolutor && !theMatBeforeUpdate )
 	currTsos = (*theConvolutor)(currTsos,
 				    usePropagator->propagationDirection());
       if(!currTsos.isValid()) {
-	edm::LogInfo("GsfTrajectorySmoother") 
+	edm::LogInfo("GsfTrackFitters") 
 	  << "GsfTrajectorySmoother: tsos not valid after update / material effects!";
 	return Trajectory();
       }
@@ -180,7 +153,7 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
       //3: combine bwd-prediction with fwd-filter
       TSOS combTsos = combiner(predTsos, (*itm).forwardPredictedState());
       if(!combTsos.isValid()) {
-	LogDebug("GsfTrajectorySmoother") << 
+	LogDebug("GsfTrackFitters") << 
 	  "KFTrajectorySmoother: combined tsos not valid!\n"<<
 	  "pred Tsos pos: "<<predTsos.globalPosition()<< "\n" <<
 	  "pred Tsos mom: "<<predTsos.globalMomentum()<< "\n" <<
@@ -191,7 +164,7 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
       TSOS smooTsos = combiner((*itm).updatedState(), predTsos);
 
       if(!smooTsos.isValid()) {
-	LogDebug("GsfTrajectorySmoother") <<
+	LogDebug("GsfTrackFitters") <<
 	  "KFTrajectorySmoother: smoothed tsos not valid!";
 	return Trajectory();
       }
@@ -201,21 +174,25 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
 	return Trajectory();
       }
 
-
+      auto chi2 = estimator()->estimate(combTsos, *(*itm).recHit()).second;
       myTraj.push(TM((*itm).forwardPredictedState(),
 		     predTsos,
 		     smooTsos,
 		     (*itm).recHit(),
-		     estimator()->estimate(combTsos, *(*itm).recHit()).second,
+		     chi2,
 		     theGeometry->idToLayer((*itm).recHit()->geographicalId() ) ),
 		  (*itm).estimate());
+       LogDebug("GsfTrackFitters") << "added measurement #" << hitcounter-- <<" with chi2 " << chi2;
+       dump(predTsos,"predTsos");
+       dump(smooTsos,"smooTsos");
+
     } 
     else {
       currTsos = predTsos;
       TSOS combTsos = combiner(predTsos, (*itm).forwardPredictedState());
       
       if(!combTsos.isValid()) {
-    	LogDebug("GsfTrajectorySmoother") << 
+    	LogDebug("GsfTrackFitters") << 
     	  "KFTrajectorySmoother: combined tsos not valid!";
     	return Trajectory();
       }
@@ -231,36 +208,35 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
     		     (*itm).recHit(),
 		     0.,
 		     theGeometry->idToLayer((*itm).recHit()->geographicalId()) ));
+       LogDebug("GsfTrackFitters") << "added invalid measurement #" << hitcounter--;
+       dump(predTsos,"predTsos");
+       dump(combTsos,"smooTsos");
     }
     if ( theMerger )  currTsos = theMerger->merge(currTsos);
+
+    dump(currTsos,"currTsos");
   }
 
   //last smoothed tm is last filtered
-  {
-    //     TimeMe t(*propTimer,false);
     predTsos = usePropagator->propagate(currTsos,
 				       *avtm.front().recHit()->surface());
-  }
   if ( predTsos.isValid() && theConvolutor && theMatBeforeUpdate )
     predTsos = (*theConvolutor)(predTsos,
 				usePropagator->propagationDirection());
   if(!predTsos.isValid()) {
-    edm::LogInfo("GsfTrajectorySmoother") << "GsfTrajectorySmoother: predicted tsos not valid!";
+    edm::LogInfo("GsfTrackFitters") << "GsfTrajectorySmoother: predicted tsos not valid!";
     return Trajectory();
   }
   if ( theMerger )  predTsos = theMerger->merge(predTsos);
 
   if(avtm.front().recHit()->isValid()) {
     //update
-    {
-      //       TimeMe t(*updateTimer,false);
-      currTsos = updator()->update(predTsos, *avtm.front().recHit());
-    }
+    currTsos = updator()->update(predTsos, *avtm.front().recHit());
     if ( currTsos.isValid() && theConvolutor && !theMatBeforeUpdate )
       currTsos = (*theConvolutor)(currTsos,
 				  usePropagator->propagationDirection());
     if(!currTsos.isValid()) {
-      edm::LogInfo("GsfTrajectorySmoother") 
+      edm::LogInfo("GsfTrackFitters") 
 	<< "GsfTrajectorySmoother: tsos not valid after update / material effects!";
       return Trajectory();
     }
@@ -270,14 +246,18 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
       return Trajectory();
     }
 
+    auto chi2 = estimator()->estimate(predTsos, *avtm.front().recHit()).second;
     myTraj.push(TM(avtm.front().forwardPredictedState(),
 		   predTsos,
 		   currTsos,
 		   avtm.front().recHit(),
-		   estimator()->estimate(predTsos, *avtm.front().recHit()).second,
+		   chi2,
 		   theGeometry->idToLayer(avtm.front().recHit()->geographicalId() )),
 		avtm.front().estimate());
-    //estimator()->estimate(predTsos, avtm.front().recHit()));
+    LogDebug("GsfTrackFitters") << "added measurement #" << hitcounter-- <<" with chi2 " << chi2;
+    dump(predTsos,"predTsos");
+    dump(currTsos,"smooTsos");
+
   }
   else {
     if (!avtm.front().forwardPredictedState().isValid()){
@@ -288,6 +268,7 @@ GsfTrajectorySmoother::trajectory(const Trajectory& aTraj) const {
 		   avtm.front().recHit(),
 		   0.,
 		   theGeometry->idToLayer(avtm.front().recHit()->geographicalId()) ));
+    LogDebug("GsfTrackFitters") << "added invalid measurement #" << hitcounter--;
   }
 
   return myTraj;
